// src/app/models/tickets.ts
export interface unAssignedAgent {
    nameIntials: string;
    firstName: string;
    lastName: string;
    assignText: string;
    icon: string;
    image: string;
}

export interface assignedAgent {
    nameIntials: string;
    firstName: string;
    lastName: string;
    icon: string;
    image: string;
}